+++ examples/v2/angular/storybook/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @copilotkit-storybook/angular
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [1510f64]
+  - @copilotkitnext/core@1.53.0
+
+## 0.0.4-next.1
+
+### Patch Changes
+
+- @copilotkitnext/core@1.53.0-next.5
+
 ## 0.0.4-next.0
 
 ### Patch Changes