+++ packages/v1/a2ui-renderer/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @copilotkit/a2ui-renderer
 
+## 1.53.0
+
+### Patch Changes
+
+- Updated dependencies [1510f64]
+- Updated dependencies [f1eaa4c]
+- Updated dependencies [ac31331]
+  - @copilotkit/react-core@1.53.0
+
+## 1.53.0-next.6
+
+### Patch Changes
+
+- @copilotkit/react-core@1.53.0-next.6
+
+## 1.53.0-next.5
+
+### Patch Changes
+
+- @copilotkit/react-core@1.53.0-next.5
+
+## 1.52.2-next.4
+
+### Patch Changes
+
+- Updated dependencies [ac31331]
+  - @copilotkit/react-core@1.52.2-next.4
+
+## 1.52.2-next.3
+
+### Patch Changes
+
+- Updated dependencies [1510f64]
+  - @copilotkit/react-core@1.52.2-next.3
+
 ## 1.52.2-next.2
 
 ### Patch Changes